requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.5.2
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
frozenlist==1.5.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jwt==1.3.1
multidict==6.4.3
oauthlib==3.2.2
propcache==0.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.1
yarl==1.19.0